Allow non-virtualenv python
Some admins may want to install the python packages using their
system rather than virtualenv.  We can remove any hard-coding to
directory paths in that case (if we assume wsgi.py is sitting in the
base dir).

We can't do that cleanly with virtualenv, since we need to load the os
or sys module in order to figure out where we were executed from.